internal/revert: Move from incusd/revert
